+++ Source/storm/storm_net.cpp
@@ -37,7 +37,7 @@ void SErrSetLastError(uint32_t dwErrCode)
 	dwLastError = dwErrCode;
 }
 
-bool SNetReceiveMessage(int *senderplayerid, void **data, uint32_t *databytes)
+bool SNetReceiveMessage(uint8_t *senderplayerid, void **data, uint32_t *databytes)
 {
 #ifndef NONET
 	std::lock_guard<SdlMutex> lg(storm_net_mutex);
